feat: Add from_unixtime support
