+++ spark/src/main/scala/org/apache/comet/serde/QueryPlanSerde.scala
@@ -31,7 +31,7 @@ import org.apache.spark.sql.catalyst.expressions.objects.StaticInvoke
 import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, NormalizeNaNAndZero}
 import org.apache.spark.sql.catalyst.plans._
 import org.apache.spark.sql.catalyst.plans.physical._
-import org.apache.spark.sql.catalyst.util.{CharVarcharCodegenUtils, TimestampFormatter}
+import org.apache.spark.sql.catalyst.util.CharVarcharCodegenUtils
 import org.apache.spark.sql.catalyst.util.ResolveDefaultColumns.getExistenceDefaultValues
 import org.apache.spark.sql.comet._
 import org.apache.spark.sql.comet.execution.shuffle.CometShuffleExchangeExec
@@ -82,6 +82,7 @@ object QueryPlanSerde extends Logging with CometExprShim {
     classOf[Chr] -> CometChr,
     classOf[InitCap] -> CometInitCap,
     classOf[BitLength] -> CometBitLength,
+    classOf[FromUnixTime] -> CometFromUnixTime,
     classOf[Length] -> CometLength,
     classOf[StringInstr] -> CometStringInstr,
     classOf[StringRepeat] -> CometStringRepeat,
@@ -1178,29 +1179,6 @@ object QueryPlanSerde extends Logging with CometExprShim {
               .build()
           })
         optExprWithInfo(optExpr, expr, child)
-
-      case FromUnixTime(sec, format, timeZoneId) =>
-        val secExpr = exprToProtoInternal(sec, inputs, binding)
-        val formatExpr = exprToProtoInternal(format, inputs, binding)
-        val timeZone = exprToProtoInternal(Literal(timeZoneId.orNull), inputs, binding)
-
-        // TODO: DataFusion toChar does not support Spark format
-        // https://github.com/apache/datafusion/issues/16577
-        // https://github.com/apache/datafusion/issues/14536
-        // TODO: DataFusion supports only -8334601211038 <= sec <= 8210266876799
-        // https://github.com/apache/datafusion/issues/16594
-        // if (secExpr.isDefined && formatExpr.isDefined) {
-        //   val timestampExpr = scalarFunctionExprToProto("from_unixtime", Seq(secExpr, timeZone): _*)
-        //   val optExpr = scalarFunctionExprToProto("to_char", Seq(timestampExpr, formatExpr): _*)
-        //   optExprWithInfo(optExpr, expr, sec, format)
-        if (secExpr.isDefined && formatExpr.isDefined && format == Literal(
-            TimestampFormatter.defaultPattern)) {
-          val optExpr = scalarFunctionExprToProto("from_unixtime", Seq(secExpr, timeZone): _*)
-          castToProto(expr, timeZoneId, StringType, optExpr.get, CometEvalMode.LEGACY)
-        } else {
-          withInfo(expr, sec, format)
-          None
-        }
 
       case IsNull(child) =>
         createUnaryExpr(
